# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clubhaus API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teams/{team_id}/microcycles": {
      "get": {
        "tags": [
          "coaching"
        ],
        "summary": "List Microcycles",
        "operationId": "list_microcycles_api_v1_teams__team_id__microcycles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 52,
              "minimum": 1,
              "default": 12,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MicrocycleResponse"
                  },
                  "title": "Response List Microcycles Api V1 Teams  Team Id  Microcycles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "coaching"
        ],
        "summary": "Create Microcycle",
        "operationId": "create_microcycle_api_v1_teams__team_id__microcycles_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MicrocycleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicrocycleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "patch": {
        "tags": [
          "coaching"
        ],
        "summary": "Update Session",
        "operationId": "update_session_api_v1_sessions__session_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/publish": {
      "post": {
        "tags": [
          "coaching"
        ],
        "summary": "Publish Session",
        "operationId": "publish_session_api_v1_sessions__session_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teams/{team_id}/performance-imports": {
      "get": {
        "tags": [
          "coaching"
        ],
        "summary": "List Performance Imports",
        "operationId": "list_performance_imports_api_v1_teams__team_id__performance_imports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerformanceImportResponse"
                  },
                  "title": "Response List Performance Imports Api V1 Teams  Team Id  Performance Imports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "coaching"
        ],
        "summary": "Create Performance Import",
        "operationId": "create_performance_import_api_v1_teams__team_id__performance_imports_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerformanceImportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "opponent": {
            "type": "string",
            "title": "Opponent"
          },
          "competition": {
            "type": "string",
            "title": "Competition"
          },
          "kickoff_at": {
            "type": "string",
            "format": "date-time",
            "title": "Kickoff At"
          },
          "venue": {
            "type": "string",
            "title": "Venue"
          },
          "home_away": {
            "type": "string",
            "title": "Home Away"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "opponent",
          "competition",
          "kickoff_at",
          "venue",
          "home_away",
          "status"
        ],
        "title": "MatchResponse"
      },
      "MicrocycleCreate": {
        "properties": {
          "opponent": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Opponent"
          },
          "competition": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Competition"
          },
          "kickoff_at": {
            "type": "string",
            "format": "date-time",
            "title": "Kickoff At"
          },
          "venue": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Venue"
          },
          "home_away": {
            "type": "string",
            "pattern": "^(home|away|neutral)$",
            "title": "Home Away",
            "default": "home"
          },
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "title": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Title"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionInput"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Sessions"
          }
        },
        "type": "object",
        "required": [
          "opponent",
          "competition",
          "kickoff_at",
          "venue",
          "week_start",
          "title",
          "sessions"
        ],
        "title": "MicrocycleCreate"
      },
      "MicrocycleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "match": {
            "$ref": "#/components/schemas/MatchResponse"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionResponse"
            },
            "type": "array",
            "title": "Sessions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "week_start",
          "title",
          "status",
          "published_at",
          "match",
          "sessions"
        ],
        "title": "MicrocycleResponse"
      },
      "PerformanceImportCreate": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "source_vendor": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Source Vendor"
          },
          "original_filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Original Filename"
          },
          "object_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Key"
          },
          "column_mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Column Mapping"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PerformanceRowInput"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "source_vendor",
          "original_filename"
        ],
        "title": "PerformanceImportCreate"
      },
      "PerformanceImportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "source_vendor": {
            "type": "string",
            "title": "Source Vendor"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Key"
          },
          "column_mapping": {
            "additionalProperties": true,
            "type": "object",
            "title": "Column Mapping"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "matched_count": {
            "type": "integer",
            "title": "Matched Count"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "session_id",
          "source_vendor",
          "original_filename",
          "object_key",
          "column_mapping",
          "row_count",
          "matched_count",
          "status",
          "created_at"
        ],
        "title": "PerformanceImportResponse"
      },
...
      "PublicationResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "session_status": {
            "type": "string",
            "title": "Session Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "audiences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Audiences"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "session_status",
          "version",
          "audiences",
          "published_at"
        ],
        "title": "PublicationResponse"
      },
...
      "SessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "day_code": {
            "type": "string",
            "title": "Day Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          },
          "intensity": {
            "type": "string",
            "title": "Intensity"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "internal_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Notes"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/SessionBlockResponse"
            },
            "type": "array",
            "title": "Blocks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "day_code",
          "title",
          "objective",
          "scheduled_at",
          "duration_minutes",
          "intensity",
          "location",
          "status",
          "version",
          "internal_notes",
          "blocks"
        ],
        "title": "SessionResponse"
      },
      "SessionUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "objective": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "intensity": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Intensity"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 180,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "internal_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Notes"
          },
          "blocks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SessionBlockInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocks"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "SessionUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}